레시피 평점기능

+++ src/main/java/com/thedish/recipe/model/vo/Recipe.java
@@ -17,7 +17,7 @@ public class Recipe implements java.io.Serializable{
 	private java.sql.Date createdAt;		//	CREATED_AT	DATE
 	private String instructions;			//	INSTRUCTIONS	VARCHAR2(4000 BYTE)
 	private int recommendNumber; 			//	RECOMMEND_NUMBER	NUMBER
-	private int avgRating;	//	AVG_RATING	NUMBER(3,1)
+	private double avgRating;	//	AVG_RATING	NUMBER(3,1)
 	private int viewCount;		//	VIEW_COUNT	NUMBER
 	private String ingredientName;		//	INGREDIENT_NAME	VARCHAR2(2000 BYTE)
 	
@@ -56,7 +56,7 @@ public class Recipe implements java.io.Serializable{
 
 
 	public Recipe(int recipeId, String name, String recipeCategory, String description, String createBy, Date createdAt,
-			String instructions, int recommendNumber, int avgRating, int viewCount, String ingredientName,
+			String instructions, int recommendNumber, double avgRating, int viewCount, String ingredientName,
 			String imageUrl) {
 		super();
 		this.recipeId = recipeId;
@@ -125,10 +125,10 @@ public class Recipe implements java.io.Serializable{
 	public void setRecommendNumber(int recommendNumber) {
 		this.recommendNumber = recommendNumber;
 	}
-	public int getAvgRating() {
+	public double getAvgRating() {
 		return avgRating;
 	}
-	public void setAvgRating(int avgRating) {
+	public void setAvgRating(double avgRating) {
 		this.avgRating = avgRating;
 	}
 	public int getViewCount() {
